# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Acestream Scraper API",
    "description": "API for scraping and managing Acestream channels",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/channels/": {
      "get": {
        "tags": ["channels"],
        "summary": "Get Channels",
        "description": "Get all channels with optional filtering.\nSupports both skip/limit and page/page_size parameters.",
        "operationId": "get_channels_api_v1_channels__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "title": "Page" }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "title": "Page Size" }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Group"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ChannelResponse" },
                  "title": "Response Get Channels Api V1 Channels  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["channels"],
        "summary": "Create Channel",
        "description": "Create a new Acestream channel. If a channel with the same ID exists, update it.",
        "operationId": "create_channel_api_v1_channels__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/channels/{channel_id}": {
      "get": {
        "tags": ["channels"],
        "summary": "Get Channel",
        "description": "Get a specific channel by ID.",
        "operationId": "get_channel_api_v1_channels__channel_id__get",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Channel Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["channels"],
        "summary": "Update Channel",
        "description": "Update an existing channel.",
        "operationId": "update_channel_api_v1_channels__channel_id__put",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Channel Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["channels"],
        "summary": "Delete Channel",
        "description": "Delete a channel.",
        "operationId": "delete_channel_api_v1_channels__channel_id__delete",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Channel Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/epg/sources": {
      "get": {
        "tags": ["epg"],
        "summary": "Get Epg Sources",
        "description": "Get all EPG sources",
        "operationId": "get_epg_sources_api_v1_epg_sources_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/EPGSourceResponse" },
                  "title": "Response Get Epg Sources Api V1 Epg Sources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["epg"],
        "summary": "Create Epg Source",
        "description": "Create a new EPG source",
        "operationId": "create_epg_source_api_v1_epg_sources_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EPGSourceCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EPGSourceResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/epg/sources/{source_id}": {
      "get": {
        "tags": ["epg"],
        "summary": "Get Epg Source",
        "description": "Get a specific EPG source by ID",
        "operationId": "get_epg_source_api_v1_epg_sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Source Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EPGSourceResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["epg"],
        "summary": "Update Epg Source",
        "description": "Update an EPG source",
        "operationId": "update_epg_source_api_v1_epg_sources__source_id__patch",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Source Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EPGSourceUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EPGSourceResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["epg"],
        "summary": "Delete Epg Source",
        "description": "Delete an EPG source",
        "operationId": "delete_epg_source_api_v1_epg_sources__source_id__delete",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Source Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/epg/sources/{source_id}/refresh": {
      "post": {
        "tags": ["epg"],
        "summary": "Refresh Epg Source",
        "description": "Refresh EPG data for a specific source",
        "operationId": "refresh_epg_source_api_v1_epg_sources__source_id__refresh_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Source Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EPGRefreshResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/epg/channels": {
      "get": {
        "tags": ["epg"],
        "summary": "Get Epg Channels",
        "description": "Get EPG channels, optionally filtered by source ID",
        "operationId": "get_epg_channels_api_v1_epg_channels_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Source Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EPGChannelResponse"
                  },
                  "title": "Response Get Epg Channels Api V1 Epg Channels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/epg/channels/{channel_id}": {
      "get": {
        "tags": ["epg"],
        "summary": "Get Epg Channel",
        "description": "Get a specific EPG channel by ID",
        "operationId": "get_epg_channel_api_v1_epg_channels__channel_id__get",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Channel Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EPGChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/epg/channels/map": {
      "post": {
        "tags": ["epg"],
        "summary": "Map Epg Channel To Tv",
        "description": "Map an EPG channel to a TV channel",
        "operationId": "map_epg_channel_to_tv_api_v1_epg_channels_map_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EPGChannelMappingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/epg/channels/unmap": {
      "post": {
        "tags": ["epg"],
        "summary": "Unmap Epg Channel From Tv",
        "description": "Remove a mapping between an EPG channel and a TV channel",
        "operationId": "unmap_epg_channel_from_tv_api_v1_epg_channels_unmap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EPGChannelMappingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/epg/channels/{channel_id}/programs": {
      "get": {
        "tags": ["epg"],
        "summary": "Get Epg Programs",
        "description": "Get programs for an EPG channel, optionally filtered by date range",
        "operationId": "get_epg_programs_api_v1_epg_channels__channel_id__programs_get",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Channel Id" }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "End Date"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EPGProgramResponse"
                  },
                  "title": "Response Get Epg Programs Api V1 Epg Channels  Channel Id  Programs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelCreate": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "source_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Source Url"
          },
          "group": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Group"
          },
          "logo": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Logo"
          },
          "tvg_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tvg Id"
          },
          "tvg_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tvg Name"
          },
          "is_online": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Online",
            "default": true
          }
        },
        "type": "object",
        "required": ["id", "name"],
        "title": "ChannelCreate",
        "description": "Schema for channel creation"
      },
      "ChannelResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "source_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Source Url"
          },
          "group": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Group"
          },
          "logo": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Logo"
          },
          "tvg_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tvg Id"
          },
          "tvg_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tvg Name"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "is_active": { "type": "boolean", "title": "Is Active" },
          "is_online": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Online"
          },
          "last_checked": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Last Checked"
          },
          "check_error": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Check Error"
          },
          "tv_channel_id": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Tv Channel Id"
          },
          "epg_update_protected": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Epg Update Protected",
            "default": false
          }
        },
        "type": "object",
        "required": ["id", "name", "last_seen", "is_active"],
        "title": "ChannelResponse",
        "description": "Schema for channel response"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}